source_list always available as attribute for Google Home/Alexa/other integrations
Ensure source_list is always available as component attribute so it can be used with Google Assistant/Alexa integration without having to have the component actually turned on.  Right now a sync devices will not sync the input sources properly when TV is off.

I do not know anything removing this code would break and have verified this in my own installation, but please doublecheck.